pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "zxgk-court-automation"
version = "1.0.0"
description = "ZXGK Court Automation Tool - Python-based desktop application for legal case automation"
authors = [
    {name = "ZXGK Team"}
]
requires-python = ">=3.8"
dependencies = [
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    "asyncio>=3.4.3",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "ddddocr>=1.0.6",
    "onnxruntime>=1.23.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "cryptography>=41.0.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "pyinstaller>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
